requirements.txt
Flask==0.12.2
line-bot-sdk==1.8.0
requests==2.22.0
urllib3==1.25.3